created update_needs function in population.py and moved calculate_item_price from item.py to sales_ledger.py to fix circular import issue
