+++ src/dataclasses/employee.py
@@ -13,6 +13,8 @@ class Employee:
     name: str
     salary: int  # how much does this employee earn in one year
     morale: float  # used to change efficiency
-    tags: List[
-        Tuple[str, int]
-    ]  # the list of skill tags where each one has an efficiency modifier between 0 and 100
+    """
+    tags is the list of skill tags where each one has an
+    efficiency modifier between 0 and 100
+    """
+    tags: List[Tuple[str, int]]
