+++ src/dataclasses/item.py
@@ -1,9 +1,6 @@
 import dataclasses as dc
-from typing import List
 
 import numpy as np
-import population as pl
-import sales_ledger as sl
 from numpy import ndarray
 
 
@@ -32,26 +29,3 @@ def item_similarity(i1: Item, i2: Item):
     dot = np.dot(i1.meaning, i2.meaning)
     norms = np.linalg.norm(i1.meaning) * np.linalg.norm(i2.meaning)
     return dot / norms  # cosine similarity
-
-
-def calculate_item_price(
-    population: pl.Population, sale_history: List[sl.SalesLedger]
-) -> float:
-    """
-    INPUT: population is the population you are selling to
-    and sales_history is a list of sales ledgers which
-    document sales of this item to this population
-    RETURNS: a price for the item
-    ERROR: none
-    SIDE-EFFECTS: none
-    This function calculates the price an item should be sold at.
-    This is a barebones placeholder implementation.
-    """
-
-    if len(sale_history) == 0:
-        return 100
-
-    avg_selling_price = sum(sale.price for sale in sale_history) / len(sale_history)
-    demand: float = float(np.linalg.norm(population.needs_state))
-
-    return avg_selling_price * demand
